tentei debugar onde tava dando erro, mas não deu ainda
